fix(0.30): conftest imports no numpy, so the wheel test still runs
CIBW_TEST_COMMAND runs tests/test_stubs.py in an environment that has
bazalt and nothing else, and a top-level `import numpy` in conftest.py
breaks collection there. The 0.30 DataType sweep added one: it inserted
the import into every file that gained a dtype= argument.

The argument was redundant anyway. All of these are INDEX buffers, and an
int list into one already infers uint32 - which is the rule 0.30 wrote
down. Dropping the dtype= drops the import with it, and the simplest
examples stop depending on numpy for one keyword.

+++ examples/02_triangle/main.py
@@ -1,5 +1,4 @@
 import bazalt as bz
-import numpy as np
 
 # Create window, logger, and renderer
 logger = bz.Logger()
@@ -31,7 +30,7 @@ vertices = [
 vbuf = ctx.create_buffer(vertices, bz.BufferUsage.VERTEX, bz.MemoryUsage.STATIC)
 
 indices = [0, 1, 2]
-ibuf = ctx.create_buffer(indices, bz.BufferUsage.INDEX, bz.MemoryUsage.STATIC, dtype=np.uint32)
+ibuf = ctx.create_buffer(indices, bz.BufferUsage.INDEX, bz.MemoryUsage.STATIC)
 
 # Build the graph once
 g = ctx.graph()
