+++ tests/conftest.py
@@ -18,7 +18,6 @@ os.environ.setdefault("BAZALT_HOT_RELOAD_POLL_MS", "20")
 import pytest
 
 import bazalt as bz
-import numpy as np
 
 SHADER_DIR = pathlib.Path(__file__).parent / "shaders"
 
@@ -209,6 +208,5 @@ def triangle_buffers(ctx):
         +0.5, +0.5, 0.0, 0.0, 0.0, 1.0,
     ]
     vbuf = ctx.create_buffer(vertices, bz.BufferUsage.VERTEX, bz.MemoryUsage.STATIC)
-    ibuf = ctx.create_buffer([0, 1, 2], bz.BufferUsage.INDEX, bz.MemoryUsage.STATIC,
-                             dtype=np.uint32)
+    ibuf = ctx.create_buffer([0, 1, 2], bz.BufferUsage.INDEX, bz.MemoryUsage.STATIC)
     return vbuf, ibuf
